+++ mrViewer/lib/mrvFl/mrvCallbacks.h
@@ -74,6 +74,9 @@ namespace mrv
     void toggle_secondary_cb( Fl_Menu_* w, ViewerUI* ui );
     void toggle_secondary_float_on_top_cb( Fl_Menu_* w, ViewerUI* ui );
 
+    //! Panel callbacks
+    void toggle_one_panel_only_cb( Fl_Menu_* w, ViewerUI* ui );
+
     void window_cb( Fl_Menu_* w, ViewerUI* ui );
     void show_window_cb( const std::string& label, ViewerUI* ui );